# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: April Profile API
  description: |
    REST API микросервиса April Profile (профиль пользователя).
    Спецификация ведётся в репозитории (`openapi/openapi.yaml`); скорректируйте `servers` под свой gateway.
    Для AprilHub BFF (фаза 4.2) публичный путь согласован как `/admin/profile/api/v1/...`,
    при этом сервисный путь остаётся `/api/v1/...` и tenant берётся только из JWT claim.
  version: 0.2.0
  license:
    name: UNLICENSED
servers:
  - url: /admin/profile/api
    description: |
      Рекомендуемый базовый путь за AprilHub BFF (same-origin админ-маршрут).
      BFF/reverse proxy снимает префикс `/admin/profile` и проксирует в AprilProfile `/api`.
  - url: https://dev.profile.april.ukituki.tech/api
    description: |
      Прямой URL сервиса на dev-стенде (без BFF Hub), используется для backend smoke.
      Подставьте базовый URL за gateway. См. `docs/guides/FORK_AND_CUSTOMIZE.md`.
...
paths:
...
  /v1/auth/whoami:
    get:
      summary: Идентичность и tenant из JWT (Keycloak)
      description: |
        `tenant_id` берётся только из согласованного claim access token (см. README / задачу 005), не из query/body.
      operationId: getAuthWhoAmI
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Текущий субъект и tenant из токена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WhoAmI'
        '401':
          description: Нет или невалидный Bearer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: В токене нет claim с tenant_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/entity-types:
    get:
      summary: Список типов сущностей tenant
      operationId: listEntityTypes
      tags: [EntityTypes]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Каталог типов сущностей текущего tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityTypeListResponse'
        '401':
          description: Нет или невалидный Bearer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: В токене нет claim с tenant_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Создать черновик типа сущности
      operationId: createEntityTypeDraft
      tags: [EntityTypes]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEntityTypeDraftRequest'
      responses:
        '201':
          description: Черновик типа создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityType'
        '400':
          description: Невалидный body или пустые поля namespace/code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Нет или невалидный Bearer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: В токене нет claim с tenant_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/entity-types/{entityTypeID}/publish:
    post:
      summary: Опубликовать черновик типа сущности (создаёт новую immutable-ревизию схемы; повторные вызовы допускаются)
      operationId: publishEntityTypeDraft
      tags: [EntityTypes]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: entityTypeID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Тип успешно опубликован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityType'
        '404':
          description: Тип не найден в текущем tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Черновик не проходит структурную валидацию/invariants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Нет или невалидный Bearer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: В токене нет claim с tenant_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/entities:
    get:
      summary: Список текущих профилей (server-side search/filter/pagination)
      operationId: listEntityProfiles
      tags: [Profiles]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: search
          required: false
          schema:
            type: string
          description: Поиск по `entity_id` и текстовым фрагментам текущего `document`.
        - in: query
          name: entity_type_id
          required: false
          schema:
            type: string
            format: uuid
          description: Фильтр по типу сущности.
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Размер страницы.
        - in: query
          name: cursor
          required: false
          schema:
            type: string
          description: Непрозрачный курсор следующей страницы.
        - in: query
          name: sort
          required: false
          schema:
            type: string
            enum: [updated_desc, updated_asc]
            default: updated_desc
          description: Детерминированная сортировка по `created_at` текущей версии + `entity_id`.
      responses:
        '200':
          description: Страница списка профилей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileListResponse'
        '401':
          description: Нет или невалидный Bearer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: В токене нет claim с tenant_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Невалидный `cursor`/`limit`/query-параметры
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Слишком много запросов (rate limit, если включён)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Создать профиль сущности (version=1)
      operationId: createEntityProfile
      tags: [Profiles]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEntityRequest'
      responses:
        '201':
          description: Профиль создан, первая версия сохранена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileSnapshot'
        '400':
          description: Невалидный payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Тип сущности не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Тип не опубликован, конфликт external mapping или все поля заблокированы authority
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Нет или невалидный Bearer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: В токене нет claim с tenant_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v1/entities/{entityID}/versions/{version}:
    get:
      summary: Получить конкретную версию профиля
      description: |
        Как и `GET /v1/entities/{entityID}` — при ABAC выдача `document` может быть усечена по ролям.
      operationId: getEntityProfileByVersion
      tags: [Profiles]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EntityID'
        - in: path
          name: version
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Конкретная версия профиля
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileSnapshot'
        '404':
          description: Версия или сущность не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Нет или невалидный Bearer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: В токене нет claim с tenant_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token Keycloak (проверка JWKS, iss, aud|azp).
        `tenant_id` извлекается только из trusted JWT claim, не из query/body/forwarded headers.
  schemas:
    WhoAmI:
      type: object
      required: [sub, tenant_id]
      properties:
        sub:
          type: string
          description: Subject из JWT (`sub`).
        tenant_id:
          type: string
          description: Идентификатор тенанта из доверенного claim (по умолчанию `tenant_id`).
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Стабильный код ошибки (машиночитаемый).
        message:
          type: string
          description: Описание для клиента и логов.
        request_id:
          description: Идентификатор запроса для корреляции в логах и трассировке.
          type: [string, 'null']
...
    CreateEntityTypeDraftRequest:
      type: object
      required: [namespace, code, draft_schema]
      properties:
        namespace:
          type: string
          example: hr
        code:
          type: string
          example: employee
        draft_schema:
          type: object
          additionalProperties: true
          example:
            type: object
            properties:
              first_name:
                type: string
              last_name:
                type: string
            required: [first_name, last_name]
...
    EntityTypeListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/EntityType'
...